# Feat: remove loop reduction & filter unnecessary apprev update