--- a/charts/vela-core/values.yaml
+++ b/charts/vela-core/values.yaml
@@ -85,7 +85,6 @@ healthCheck:
 
 ## @section KubeVela controller optimization parameters
 ##@param optimize.cachedGvks Optimize types of resources to be cached.
-##@param optimize.controllerReconcileLoopReduction Optimize ApplicationController reconcile by reducing the number of loops to reconcile application.
 ##@param optimize.markWithProb Optimize ResourceTracker GC by only run mark with probability. Side effect: outdated ResourceTracker might not be able to be removed immediately.
 ##@param optimize.disableComponentRevision Optimize componentRevision by disabling the creation and gc
 ##@param optimize.disableApplicationRevision Optimize ApplicationRevision by disabling the creation and gc.
@@ -95,7 +94,6 @@ healthCheck:
 ##@param optimize.enableResourceTrackerDeleteOnlyTrigger Optimize resourcetracker by only trigger reconcile when resourcetracker is deleted.
 optimize:
   cachedGvks: ""
-  controllerReconcileLoopReduction: false
   markWithProb: 0.1
   disableComponentRevision: true
   disableApplicationRevision: false