### Description of your changes

<!--

Briefly describe what this pull request does. We love pull requests that resolve an open KubeVela issue. If yours does, you
can uncomment the below line to indicate which issue your PR fixes, for example
"Fixes #500":

-->

Add two optimization:
1. Remove `loop-reduction` optimization and let it be enabled by default. This is because the workflow state transition has changed its mechanism and if we do not enable loop reduction by default, the application will run twice gc every reconcile.
2. The application revision updates always runs no matter whether it is needed. However, usually, the application revision do not change when state-keep-oriented reconcile happens. This PR reduct it to optimize performance.

Result:
![image](https://user-images.githubusercontent.com/14019297/222616934-32eea278-68ba-45dd-818c-19a4e435571c.png)
Optimization 1 can reduce the reconcile time from 125ms to 96ms, increase 23% performance for state-keep-oriented reconcile under simple application delivery scenario.
![image](https://user-images.githubusercontent.com/14019297/222617094-1fe1f2c6-58f8-4362-8754-5d134aed6f85.png)
Optimization 2 can reduce the reconcile time from 96ms to 70ms, increase 27% performance for state-keep oriented reconcile under simple application delivery scenario.

I have:

- [ ] Read and followed KubeVela's [contribution process](https://github.com/kubevela/kubevela/blob/master/contribute/create-pull-request.md).
- [ ] [Related Docs](https://github.com/kubevela/kubevela.io) updated properly. In a new feature or configuration option, an update to the documentation is necessary. 
- [ ] Run `make reviewable` to ensure this PR is ready for review.
- [ ] Added `backport release-x.y` labels to auto-backport this PR if necessary.

### How has this code been tested

<!--
Before reviewers can be confident in the correctness of this pull request, it
needs to tested and shown to be correct. Briefly describe the testing that has
already been done or which is planned for this change.
-->


### Special notes for your reviewer

<!--

Be sure to direct your reviewers'
attention to anything that needs special consideration.

-->